// client/src/client/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseUrl: `${string}://${string}` | (string & {});
};

/**
 * BaseRequest
 */
export type BaseRequest = {
    /**
     * Url
     */
    url: string;
    /**
     * Implicit Rows
     */
    implicit_rows?: boolean;
    /**
     * Implicit Columns
     */
    implicit_columns?: boolean;
    /**
     * Borderless Tables
     */
    borderless_tables?: boolean;
    /**
     * Min Confidence
     */
    min_confidence?: number;
    /**
     * Detect Rotation
     */
    detect_rotation?: boolean;
};

/**
 * BoundingBox
 */
export type BoundingBox = {
    /**
     * X1
     */
    x1: number;
    /**
     * Y1
     */
    y1: number;
    /**
     * X2
     */
    x2: number;
    /**
     * Y2
     */
    y2: number;
};

/**
 * CellContent
 */
export type CellContent = {
    /**
     * Content
     */
    content?: string | null;
};

/**
 * HTTPValidationError
 */
export type HttpValidationError = {
    /**
     * Detail
     */
    detail?: Array<ValidationError>;
};

/**
 * PDFRequest
 */
export type PdfRequest = {
    /**
     * Url
     */
    url: string;
    /**
     * Implicit Rows
     */
    implicit_rows?: boolean;
    /**
     * Implicit Columns
     */
    implicit_columns?: boolean;
    /**
     * Borderless Tables
     */
    borderless_tables?: boolean;
    /**
     * Min Confidence
     */
    min_confidence?: number;
    /**
     * Detect Rotation
     */
    detect_rotation?: boolean;
    /**
     * Pages
     */
    pages?: Array<number> | null;
    /**
     * Pdf Text Extraction
     */
    pdf_text_extraction?: boolean;
};

/**
 * PdfTableResponse
 */
export type PdfTableResponse = {
    box: BoundingBox;
    /**
     * Title
     */
    title: string | null;
    /**
     * Grid
     */
    grid: Array<Array<number>>;
    /**
     * Cells
     */
    cells: Array<CellContent>;
    /**
     * Page
     */
    page: number;
};

/**
 * TableResponse
 */
export type TableResponse = {
    box: BoundingBox;
    /**
     * Title
     */
    title: string | null;
    /**
     * Grid
     */
    grid: Array<Array<number>>;
    /**
     * Cells
     */
    cells: Array<CellContent>;
};

/**
 * ValidationError
 */
export type ValidationError = {
    /**
     * Location
     */
    loc: Array<string | number>;
    /**
     * Message
     */
    msg: string;
    /**
     * Error Type
     */
    type: string;
};

export type ProcessImageImagePostData = {
    body: BaseRequest;
    path?: never;
    query?: never;
    url: '/image';
};

export type ProcessImageImagePostErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type ProcessImageImagePostError = ProcessImageImagePostErrors[keyof ProcessImageImagePostErrors];

export type ProcessImageImagePostResponses = {
    /**
     * Response Process Image Image Post
     *
     * Successful Response
     */
    200: Array<TableResponse>;
};

export type ProcessImageImagePostResponse = ProcessImageImagePostResponses[keyof ProcessImageImagePostResponses];

export type ProcessPdfPdfPostData = {
    body: PdfRequest;
    path?: never;
    query?: never;
    url: '/pdf';
};

export type ProcessPdfPdfPostErrors = {
    /**
     * Validation Error
     */
    422: HttpValidationError;
};

export type ProcessPdfPdfPostError = ProcessPdfPdfPostErrors[keyof ProcessPdfPdfPostErrors];

export type ProcessPdfPdfPostResponses = {
    /**
     * Response Process Pdf Pdf Post
     *
     * Successful Response
     */
    200: Array<PdfTableResponse>;
};

export type ProcessPdfPdfPostResponse = ProcessPdfPdfPostResponses[keyof ProcessPdfPdfPostResponses];